+++ src/components/mbx-button/index.js
@@ -1,3 +1,0 @@
-import main from './mbx-button';
-
-export default main;