+++ src/components/mbx-icon-text/index.js
@@ -1,3 +1,0 @@
-import main from './mbx-icon-text';
-
-export default main;