+++ src/components/mbx-underline-tabs/index.js
@@ -1,3 +1,0 @@
-import main from './mbx-underline-tabs';
-
-export default main;